feature: ADD save image controller layer.
