+++ sse_api/src/apps/__init__.py
@@ -1,0 +1,24 @@
+from fastapi import FastAPI
+from fastapi.middleware.cors import CORSMiddleware
+from .controller import user
+from src.libs.error_handler import error_handlers
+
+
+def create_app():
+    app = FastAPI()
+
+    # Router
+    app.include_router(user)
+
+    # Handler
+    error_handlers(app)
+
+    # CORS
+    app.add_middleware(
+        CORSMiddleware,
+        allow_origins=["*"],
+        allow_credentials=True,
+        allow_methods=["*"],
+        allow_headers=["*"],
+    )
+    return app
